created reusable tabs + breadcrums components (#896)
* created reusable tabs + breadcrums components

Looks like I changed a lot of files, but I didn't. I just moved files and grouped all the logic related to the contracts in the contracts file.

* DEV-452-Network-Configs (#897)

* load networks from firebase

* add networks routes

* add networks deploy

* fix types router

* fix create2factory tx for networks

* DEV-453-Refactor-Crud-Firebase (#899)

* update firebase web crud

* refactor db paths as vars

* refactor crud web exports

* refactor admin crud

* refactor hooks exports

* crud-firebase

* fix contracts-api-firebase

* fix tests composite id

* rename ItemComposite

* fix hooks

* update models

* fix validators and models

* update models

* fix validators for contract

* fix new firebase crud integration for contracts-api

* fix dashboard build, zod-sol isolate node dep

* fix lint

* created reusable tabs + breadcrums components

Looks like I changed a lot of files, but I didn't. I just moved files and grouped all the logic related to the contracts in the contracts file.

* Update index.tsx

* DEV-419-EthModels-in-api (#900)

track events and transactions for codegen contracts

* [WIP] modified welcome email

* created reusable tabs + breadcrums components

Looks like I changed a lot of files, but I didn't. I just moved files and grouped all the logic related to the contracts in the contracts file.

* created reusable tabs + breadcrums components

Looks like I changed a lot of files, but I didn't. I just moved files and grouped all the logic related to the contracts in the contracts file.

* [WIP] modified welcome email

* change localhost to 127.0.0.1

* Removed dummy data and components

* added tables for events and transactions + some refactoring

* created reusable tabs + breadcrums components

Looks like I changed a lot of files, but I didn't. I just moved files and grouped all the logic related to the contracts in the contracts file.

* [WIP] modified welcome email

* change localhost to 127.0.0.1

* Removed dummy data and components

* added tables for events and transactions + some refactoring

* skip test erc20 + fices to events table

---------

Co-authored-by: Leo Vigna <leo.vigna@gmail.com>